Revert "Refactor agent controller to switch between http and websockets"
