+++ agent-common/test/com/thoughtworks/go/agent/launcher/ServerBinaryDownloaderTest.java
@@ -16,7 +16,6 @@
 
 package com.thoughtworks.go.agent.launcher;
 
-import com.thoughtworks.go.agent.common.ssl.GoAgentServerHttpClientBuilder;
 import com.thoughtworks.go.agent.common.util.Downloader;
 import com.thoughtworks.go.agent.testhelper.FakeBootstrapperServer;
 import com.thoughtworks.go.mothers.ServerUrlGeneratorMother;
@@ -118,7 +117,7 @@ public class ServerBinaryDownloaderTest {
     @Test
     public void shouldFailIfServerIsNotAvailable() throws Exception {
         exception.expect(UnknownHostException.class);
-        exception.expectMessage("invalidserver: Name or service not known");
+        exception.expectMessage("invalidserver: unknown error");
 
         ServerBinaryDownloader downloader = new ServerBinaryDownloader(ServerUrlGeneratorMother.generatorWithoutSubPathFor("https://invalidserver:9091/go/hello"), null,
                 SslVerificationMode.NONE);
@@ -138,7 +137,7 @@ public class ServerBinaryDownloaderTest {
 
     @Test
     public void shouldReturnFalseIfTheServerDoesNotRespondWithEntity() throws Exception {
-        GoAgentServerHttpClientBuilder builder = mock(GoAgentServerHttpClientBuilder.class);
+        HttpClientBuilder builder = mock(HttpClientBuilder.class);
         CloseableHttpClient closeableHttpClient = mock(CloseableHttpClient.class);
         when(builder.build()).thenReturn(closeableHttpClient);
         CloseableHttpResponse httpResponse = mock(CloseableHttpResponse.class);
