+++ agent/build.gradle
@@ -31,9 +31,8 @@ dependencies {
   compile project(':common')
   compile group: 'org.objenesis', name: 'objenesis', version: '1.2'
   compile group: 'org.bouncycastle', name: 'bcprov-jdk15on', version: '1.47'
-  compile group: 'commons-configuration', name: 'commons-configuration', version: '1.10'
   compile group: 'org.eclipse.jetty.websocket', name: 'websocket-client', version: versions.jetty
-
+  compile group: 'commons-configuration', name: 'commons-configuration', version: '1.10'
   testCompile project(path: ':common', configuration: 'testOutput')
   testCompile project(path: ':config-api', configuration: 'testOutput')
   testCompile project(':test-utils')
