+++ agent/resources/agent.properties
@@ -15,5 +15,8 @@
 ##########################GO-LICENSE-END##################################
 
 agent.instruction.interval=10000
+agent.instruction.delay=1500
 agent.ping.interval=5000
+agent.ping.delay=1500
 agent.get.work.interval=10000
+agent.get.work.delay=1000
