+++ agent/test/functional/com/thoughtworks/go/agent/functional/AgentStatusReportingFunctionalTest.java
@@ -16,7 +16,7 @@
 
 package com.thoughtworks.go.agent.functional;
 
-import com.thoughtworks.go.agent.common.ssl.GoAgentServerClientBuilder;
+import com.thoughtworks.go.agent.common.ssl.GoAgentServerHttpClientBuilder;
 import com.thoughtworks.go.agent.testhelpers.FakeBuildRepositoryRemote;
 import com.thoughtworks.go.agent.testhelpers.FakeGoServer;
 import com.thoughtworks.go.agent.testhelpers.LongWorkCreator;
@@ -51,8 +51,8 @@ public class AgentStatusReportingFunctionalTest {
     public static void systemProperties() throws Exception {
         new SystemEnvironment().setProperty("serviceUrl", SERVER_URL);
         new SystemEnvironment().setProperty("WORKCREATOR", LongWorkCreator.class.getCanonicalName());
-        GoAgentServerClientBuilder.AGENT_CERTIFICATE_FILE.delete();
-        GoAgentServerClientBuilder.AGENT_TRUST_FILE.delete();
+        GoAgentServerHttpClientBuilder.AGENT_CERTIFICATE_FILE.delete();
+        GoAgentServerHttpClientBuilder.AGENT_TRUST_FILE.delete();
         fakeGoServer = new FakeGoServer(9090, 8443);
         fakeGoServer.start();
         FakeBuildRepositoryRemote.AGENT_STATUS.clear();
