+++ agent/test/functional/com/thoughtworks/go/agent/service/SslInfrastructureServiceTest.java
@@ -17,7 +17,6 @@
 package com.thoughtworks.go.agent.service;
 
 import com.thoughtworks.go.agent.AgentAutoRegistrationPropertiesImpl;
-import com.thoughtworks.go.agent.common.ssl.GoAgentServerClientBuilder;
 import com.thoughtworks.go.agent.common.ssl.GoAgentServerHttpClient;
 import com.thoughtworks.go.agent.common.ssl.GoAgentServerHttpClientBuilder;
 import com.thoughtworks.go.agent.testhelpers.AgentCertificateMother;
@@ -58,8 +57,8 @@ public class SslInfrastructureServiceTest {
     @Before
     public void setup() throws Exception {
         remoteCalled = false;
-        GoAgentServerClientBuilder.AGENT_CERTIFICATE_FILE.delete();
-        GoAgentServerClientBuilder.AGENT_CERTIFICATE_FILE.deleteOnExit();
+        GoAgentServerHttpClientBuilder.AGENT_CERTIFICATE_FILE.delete();
+        GoAgentServerHttpClientBuilder.AGENT_CERTIFICATE_FILE.deleteOnExit();
         Registration registration = createRegistration();
         sslInfrastructureService = new SslInfrastructureService(requesterStub(registration), httpClientStub());
         GuidService.storeGuid("uuid");
@@ -68,8 +67,8 @@ public class SslInfrastructureServiceTest {
     @After
     public void teardown() throws Exception {
         GuidService.deleteGuid();
-        GoAgentServerClientBuilder.AGENT_CERTIFICATE_FILE.delete();
-        GoAgentServerClientBuilder.AGENT_TRUST_FILE.delete();
+        GoAgentServerHttpClientBuilder.AGENT_CERTIFICATE_FILE.delete();
+        GoAgentServerHttpClientBuilder.AGENT_TRUST_FILE.delete();
     }
 
     @Test
@@ -80,7 +79,7 @@ public class SslInfrastructureServiceTest {
         shouldCreateSslInfrastucture();
 
         sslInfrastructureService.registerIfNecessary(new AgentAutoRegistrationPropertiesImpl(configFile));
-        assertThat(GoAgentServerClientBuilder.AGENT_CERTIFICATE_FILE, exists());
+        assertThat(GoAgentServerHttpClientBuilder.AGENT_CERTIFICATE_FILE, exists());
         assertRemoteCalled();
 
         sslInfrastructureService.registerIfNecessary(new AgentAutoRegistrationPropertiesImpl(configFile));
@@ -114,8 +113,7 @@ public class SslInfrastructureServiceTest {
 
     private SslInfrastructureService.RemoteRegistrationRequester requesterStub(final Registration registration) {
         final SslInfrastructureServiceTest me = this;
-        final SystemEnvironment systemEnvironment = new SystemEnvironment();
-        return new SslInfrastructureService.RemoteRegistrationRequester(null, agentRegistryStub(), new GoAgentServerHttpClient(new GoAgentServerHttpClientBuilder(systemEnvironment))) {
+        return new SslInfrastructureService.RemoteRegistrationRequester(null, agentRegistryStub(), new GoAgentServerHttpClient(new SystemEnvironment())) {
             protected Registration requestRegistration(String agentHostName, AgentAutoRegistrationProperties agentAutoRegisterProperties)
                     throws IOException, ClassNotFoundException {
                 LOGGER.debug("Requesting remote registration");
