+++ base/src/com/thoughtworks/go/util/SystemEnvironment.java
@@ -779,13 +779,6 @@ public class SystemEnvironment implements Serializable, ConfigDirProvider {
         return GO_AGENT_USE_SSL_CONTEXT.getValue();
     }
 
-    public boolean isBuildCommandProtocolEnabled() {
-        return Boolean.valueOf(get(SystemEnvironment.ENABLE_BUILD_COMMAND_PROTOCOL));
-    }
-
-    public String getAgentKeyStorePassword() {
-        return get(SystemEnvironment.GO_AGENT_KEYSTORE_PASSWORD);
-    }
 
     public static abstract class GoSystemProperty<T> {
         private String propertyName;
