+++ common/src/com/thoughtworks/go/config/AgentAutoRegistrationProperties.java
@@ -30,6 +30,4 @@ public interface AgentAutoRegistrationProperties {
     String agentAutoRegisterHostname();
 
     void scrubRegistrationProperties();
-
-    boolean isElastic();
 }
