+++ common/src/com/thoughtworks/go/websocket/Action.java
@@ -21,5 +21,5 @@ public enum Action {
     cancelBuild,
     ping,
     reregister,
-    reportCurrentStatus, reportCompleted, reportCompleting, acknowledge, build, setCookie
+    reportCurrentStatus, reportCompleted, reportCompleting, ack, build, setCookie
 }
