+++ common/src/com/thoughtworks/go/websocket/Message.java
@@ -27,17 +27,22 @@ public class Message {
     @Expose
     private final String data;
     @Expose
-    private final String acknowledgementId;
+    private String ackId;
 
     public Message(Action action) {
         this(action, null);
     }
 
     public Message(Action action, String data) {
+        this(action, data, null);
+    }
+
+    public Message(Action action, String data, String ackId) {
         this.action = action;
         this.data = data;
-        this.acknowledgementId = UUID.randomUUID().toString();
+        this.ackId = ackId;
     }
+
 
     public Action getAction() {
         return action;
@@ -47,8 +52,8 @@ public class Message {
         return data;
     }
 
-    public String getAcknowledgementId() {
-        return acknowledgementId;
+    public String getAckId() {
+        return ackId;
     }
 
     @Override
@@ -56,7 +61,7 @@ public class Message {
         return "Message{" +
                 "action=" + action +
                 ", data=" + data +
-                ", acknowledgementId=" + acknowledgementId +
+                ", ackId=" + ackId +
                 '}';
     }
 
@@ -69,7 +74,7 @@ public class Message {
 
         if (action != message.action) return false;
         if (data != null ? !data.equals(message.data) : message.data != null) return false;
-        return acknowledgementId != null ? acknowledgementId.equals(message.acknowledgementId) : message.acknowledgementId == null;
+        return ackId != null ? ackId.equals(message.ackId) : message.ackId == null;
 
     }
 
@@ -77,8 +82,11 @@ public class Message {
     public int hashCode() {
         int result = action.hashCode();
         result = 31 * result + (data != null ? data.hashCode() : 0);
-        result = 31 * result + (acknowledgementId != null ? acknowledgementId.hashCode() : 0);
+        result = 31 * result + (ackId != null ? ackId.hashCode() : 0);
         return result;
     }
 
+    public void generateAckId() {
+        this.ackId = UUID.randomUUID().toString();
+    }
 }
